requirements.txt
aiomysql==0.2.0
fastapi==0.115.0
passlib==1.7.4
pydantic==2.9.2
python_jose==3.3.0
uvicorn==0.15.0
